pyproject.toml
[tool.poetry]
name = "rest-template"
version = "0.1.0"
description = ""
authors = ["CoSm1C AvEnGeR <zhora.zhilin.06@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.115.4"}
asyncpg = "^0.30.0"
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
python-dotenv = "^1.0.1"
redis = "^5.2.0"
python-multipart = "^0.0.17"
pyjwt = {extras = ["crypto"], version = "^2.9.0"}
dishka = "^1.4.0"
argon2-cffi = "^23.1.0"
gunicorn = "^23.0.0"
pytz = "^2024.2"


[tool.poetry.group.dev.dependencies]
setuptools = "^75.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
